+++ websocket_server.py
@@ -1,8 +1,9 @@
 import socket, base64, hashlib, json, time
 from logger import log_it
-from config import Config, Globals
+from config import Globals
 from thread_launcher import start_thread
 from utils import is_port_available
+from datetime import datetime
 
 def handshake(conn):
     request = conn.recv(1024).decode()