+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
-## [5.03] - 2025-06-09
+## [5.10] - 2025-06-15
 
 ### Fixed
-- Reset heartbeat "in-node-list" notifications to 0 when everything is OK.
-- Text indentation in notification when node is not in node list.
+- Missing `<title>` on WebUI.
+- Infinite "node-in-node-list" notifications.
+- Some other strange things which were happening with `Heartbeat`.
+
+### Added
+- Support for interacting with Cellframe Node CLI.